+++ cmd/npv/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/cybozu-go/network-policy-viewer/cmd/npv/sub"
+
+func main() {
+	sub.Execute()
+}
