+++ cmd/npv/sub/dump.go
@@ -1,4 +1,4 @@
-package cmd
+package sub
 
 import (
 	"bytes"
