+++ cmd/npv/sub/helper.go
@@ -1,4 +1,4 @@
-package cmd
+package sub
 
 import (
 	"context"
