+++ cmd/npv/sub/list.go
@@ -1,4 +1,4 @@
-package cmd
+package sub
 
 import (
 	"context"
