+++ cmd/npv/sub/root.go
@@ -1,4 +1,4 @@
-package cmd
+package sub
 
 import (
 	"fmt"
