--- main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/cybozu-go/network-policy-viewer/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
